* master:
  rw_spinlock: constructor fix

+++ nova-tt/rw_spinlock.hpp
@@ -80,7 +80,7 @@ public:
     };
 
     rw_spinlock(void):
-        state(unlocked_state)
+        state(uint32_t(unlocked_state))
     {}
 
     ~rw_spinlock(void)
